requirements.txt
numpy~=1.26.4
pillow~=10.2.0
datasets~=2.16.1
matplotlib~=3.8.2
torch~=2.2.0
ipython~=8.21.0
tqdm~=4.66.1
pandas~=2.2.0
scikit-learn~=1.4.0
torchvision~=0.17.0
